+++ modules/language/inventory/DictionaryCreatePopup.tsx
@@ -2,12 +2,12 @@ import { Button, Input, Modal } from '@/components'
 import { apiRoute } from '@/constants/apiRoutes'
 import { TOKEN_AUTHENTICATION, USER_ID } from '@/constants/auth'
 import { useApiCall, useTranslation, useTranslationFunction } from '@/hooks'
+import { inputStyles } from '@/inventory'
 import { postMethod } from '@/services'
 import { DictionaryKey } from '@/types'
 import { useEffect, useState } from 'react'
 import { useCookies } from 'react-cookie'
 import { toast } from 'react-toastify'
-import { inputStylesLanguage } from './Language.inventory'
 
 interface IDictionaryCreatePopup {
   listKeyOfDictionary: string[]
@@ -95,7 +95,7 @@ export const DictionaryCreatePopup = ({
                   [key]: event.currentTarget.value,
                 })
               }}
-              {...inputStylesLanguage({
+              {...inputStyles({
                 error: createResult?.error?.result.key && translate(createResult.error.result.key),
               })}
             />
