+++ modules/paths/inventory/DeletePathPopup.tsx
