--- modules/per-authentication/verify/verify.inventory.tsx
@@ -1,12 +1,0 @@
-export const inputStyles = ({ error }: { error?: string }) => {
-  const initialValue: Partial<any> = {
-    clearable: true,
-    status: error ? 'error' : undefined,
-    helperText: error || '',
-    underlined: true,
-  }
-
-  return {
-    ...initialValue,
-  }
-}
