+++ modules/permission/detail/PermissionDetail.tsx
@@ -2,6 +2,7 @@ import { Button, Loading } from '@/components'
 import { apiRoute } from '@/constants/apiRoutes'
 import { TOKEN_AUTHENTICATION } from '@/constants/auth'
 import { useApiCall, useGetBreadCrumb, useTranslation, useTranslationFunction } from '@/hooks'
+import { PermissionResponseDefault } from '@/inventory'
 import { getListEditAble, lostOddProps } from '@/lib'
 import { ShareStoreSelector } from '@/redux/share-store'
 import { getMethod, putMethod } from '@/services'
@@ -11,7 +12,7 @@ import { useEffect, useState } from 'react'
 import { useCookies } from 'react-cookie'
 import { useSelector } from 'react-redux'
 import { toast } from 'react-toastify'
-import { ModifierPermission, PermissionResponseDefault } from '../inventory'
+import { ModifierPermission } from '../inventory'
 import { DeletePermissionPopup } from '../inventory/DeletePermissionPopup'
 
 export const PermissionDetail = () => {
