+++ modules/permission/inventory/index.tsx
@@ -1,3 +1,1 @@
 export * from './ModifierPermission'
-export * from './UserTable'
-export * from './permission.inventory'
