+++ modules/settings/inventory/UpdateAccount.tsx
@@ -2,8 +2,8 @@ import { Button, Loading } from '@/components'
 import { apiRoute } from '@/constants/apiRoutes'
 import { TOKEN_AUTHENTICATION, USER_ID } from '@/constants/auth'
 import { useApiCall, useTranslation, useTranslationFunction } from '@/hooks'
+import { DefaultUser, UserForm } from '@/inventory'
 import { getListEditAble, lostOddProps } from '@/lib'
-import { DefaultUser, UserForm } from '@/modules/user/inventory'
 import { getMethod, putMethod } from '@/services'
 import { UpdateAccountFailure, UpdateAccountRequest, UserResponseSuccess } from '@/types'
 import { useEffect, useState } from 'react'
