--- modules/user/inventory/index.tsx
@@ -1,2 +1,0 @@
-export * from './UserForm'
-export * from './User.inventory'
